perform deletion of options on delete theme action

+++ wp_delete_theme_options.php
@@ -8,20 +8,25 @@ Version: 0.1
 Author URI: http://blogdesignstudio
 */
 
-function ravs_themeDeactivationFx( $oldname, $oldtheme=false ) {
-	// update_option( 'ravs_theme', trim( str_replace( ' ', '_', strtolower( $oldname ) ) ) );
+function ravs_themeDeactivationFx( ) {
 	
-	global $wpdb;
-	$theme_name = str_replace( ' ', '_', trim( strtolower( $oldname ) ) ); // theme name
+	// check user deleting theme or not
+	if( ( isset( $_POST['action'] ) && $_POST['action'] ==='delete' ) && ( isset( $_POST['stylesheet'] ) && $_POST['stylesheet'] !='' ) ){
 	
-	$all_options_of_theme = $wpdb->get_results("SELECT option_name FROM $wpdb->options WHERE option_name LIKE  '%{$theme_name}%'"); // get all option
-
-	if( ! empty( $all_options_of_theme ) ){
-		foreach ($all_options_of_theme as $option) {
-			delete_option( $option->option_name );
+		global $wpdb;
+		$theme_name = str_replace( ' ', '_', trim( strtolower( $oldname ) ) ); // theme name
+		
+		$all_options_of_theme = $wpdb->get_results("SELECT option_name FROM $wpdb->options WHERE option_name LIKE  '%{$theme_name}%'"); // get all option
+	
+		if( ! empty( $all_options_of_theme ) ){
+			foreach ($all_options_of_theme as $option) {
+				delete_option( $option->option_name );
+			}
 		}
 	}
 }
-add_action( "after_switch_theme", "ravs_themeDeactivationFx", 10 , 2 );
+
+if( is_admin() )
+	add_action( "init", "ravs_themeDeactivationFx" );
 
 ?>
